debug

+++ include/MotionController.hpp
@@ -63,8 +63,8 @@ private:
     std::shared_ptr<long> translationSpeed = std::shared_ptr<long>(new long(0));		// ticks/seconde
 
     PID curvePID; //FIXME INIT
-    std::shared_ptr<long> curveSetpoint = std::shared_ptr<long>(new long(INT64_MAX));
-    std::shared_ptr<long> currentRadius = std::shared_ptr<long>(new long(INT64_MAX));
+    std::shared_ptr<long> curveSetpoint = std::shared_ptr<long>(new long(INT32_MAX));
+    std::shared_ptr<long> currentRadius = std::shared_ptr<long>(new long(INT32_MAX));
     std::shared_ptr<long> deltaRadius = std::shared_ptr<long>(new long(0));
 
     //	Limitation de vitesses
